Avance de la solución.

+++ mooc_node-mod10_quiz_mvc_server.js
@@ -1,70 +1,69 @@
-
 const express = require('express');
 const app = express();
 
-   // Import MW for parsing POST params in BODY
+// Import MW for parsing POST params in BODY - Necesario cuando usamos POST, para acceder a los parametros que vienen en el body.
 
 var bodyParser = require('body-parser');
-app.use(bodyParser.urlencoded({ extended: true }));
+app.use(bodyParser.urlencoded({extended: true}));
 
-   // Import MW supporting Method Override with express
+// Import MW supporting Method Override with express - Los formularios solo pueden ser GET o POST, a pesar de usar uno de esos, queremos sobreescribir el metodo para que sea un PUT, DELETE
 
 var methodOverride = require('method-override');
-app.use(methodOverride('_method'));
+app.use(methodOverride('_method', { methods: ['POST','GET']})); // <= OJO, el parametro final POST y GET para que se aplique a los dos metodos. Si no se pone, solo se aplica a POST
 
 
-   // MODEL
+// MODEL
 
 const Sequelize = require('sequelize');
 
-const options = { logging: false, operatorsAliases: false};
+const options = {logging: false, operatorsAliases: false};
 const sequelize = new Sequelize("sqlite:db.sqlite", options);
 
 const quizzes = sequelize.define(  // define table quizzes
-    'quizzes',     
-    {   question: Sequelize.STRING,
+    'quizzes',
+    {
+        question: Sequelize.STRING,
         answer: Sequelize.STRING
     }
 );
 
 sequelize.sync() // Syncronize DB and seed if needed
-.then(() => quizzes.count())
-.then((count) => {
-    if (count===0) {
-        return ( 
-            quizzes.bulkCreate([
-                { id: 1, question: "Capital of Italy",    answer: "Rome" },
-                { id: 2, question: "Capital of France",   answer: "Paris" },
-                { id: 3, question: "Capital of Spain",    answer: "Madrid" },
-                { id: 4, question: "Capital of Portugal", answer: "Lisbon" }
-            ])
-            .then( c => console.log(`  DB created with ${c.length} elems`))
-        )
-    } else {
-        return console.log(`  DB exists & has ${count} elems`);
-    }
-})
-.catch( err => console.log(`   ${err}`));
+    .then(() => quizzes.count())
+    .then((count) => {
+        if (count === 0) {
+            return (
+                quizzes.bulkCreate([
+                    {id: 1, question: "Capital of Italy", answer: "Rome"},
+                    {id: 2, question: "Capital of France", answer: "Paris"},
+                    {id: 3, question: "Capital of Spain", answer: "Madrid"},
+                    {id: 4, question: "Capital of Portugal", answer: "Lisbon"}
+                ])
+                    .then(c => console.log(`  DB created with ${c.length} elems`))
+            )
+        } else {
+            return console.log(`  DB exists & has ${count} elems`);
+        }
+    })
+    .catch(err => console.log(`   ${err}`));
 
 
-   // VIEWs
+// VIEWs
 
 const index = (quizzes) => `<!-- HTML view -->
 <html>
     <head><title>MVC Example</title><meta charset="utf-8"></head> 
     <body> 
-        <h1>MVC: Quizzes</h1>`
-+ quizzes.reduce(
-    (ac, quiz) => ac += 
-`       <a href="/quizzes/${quiz.id}/play">${quiz.question}</a>
-        <a href="/quizzes/${quiz.id}/edit"><button>Edit</button></a>
-        <a href="/quizzes/${quiz.id}?_method=DELETE"
-           onClick="return confirm('Delete: ${quiz.question}')">
-           <button>Delete</button></a>
-        <br>\n`, 
-    ""
-)
-+ `     <p/>
+        <h1>MVC: Quizzes</h1> 
+        <table border="0"> `
+    + quizzes.reduce(
+        (ac, quiz) => ac +=
+            `<tr>   <td> <a href="/quizzes/${quiz.id}/play">${quiz.question}</a> </td>
+        <td><a href="/quizzes/${quiz.id}/edit"><button>Edit</button></a> </td>
+        <td><a href="/quizzes/${quiz.id}?_method=DELETE" onClick="return confirm('Delete: ${quiz.question}')">  <button>Delete</button></a> </td>
+        <br>\n </tr>`,
+        ""
+    )
+    + `</table>     <p/>
         <a href="/quizzes/new"><button>New Quiz</button></a>
     </body>
 </html>`;
@@ -96,7 +95,7 @@ const check = (id, msg, response) => `<!-- HTML view -->
     </body>
 </html>`;
 
-const quizForm =(msg, method, action, question, answer) => `<!-- HTML view -->
+const quizForm = (msg, method, event, action, question, answer) => `<!-- HTML view -->
 <html>
     <head><title>MVC Example</title><meta charset="utf-8"></head> 
     <body>
@@ -105,7 +104,7 @@ const quizForm =(msg, method, action, question, answer) => `<!-- HTML view -->
             ${msg}: <p>
             <input  type="text"  name="question" value="${question}" placeholder="Question" />
             <input  type="text"  name="answer"   value="${answer}"   placeholder="Answer" />
-            <input  type="submit" value="Create"/> <br>
+            <input  type="submit" value="${event}"/> <br>
         </form>
         </p>
         <a href="/quizzes"><button>Go back</button></a>
@@ -113,14 +112,14 @@ const quizForm =(msg, method, action, question, answer) => `<!-- HTML view -->
 </html>`;
 
 
-   // CONTROLLER
+// CONTROLLER
 
 // GET /, GET /quizzes
 const indexController = (req, res, next) => {
- 
+
     quizzes.findAll()
-    .then((quizzes) => res.send(index(quizzes)))
-    .catch((error) => `DB Error:\n${error}`);
+        .then((quizzes) => res.send(index(quizzes)))
+        .catch((error) => `DB Error:\n${error}`);
 }
 
 //  GET  /quizzes/1/play
@@ -129,9 +128,9 @@ const playController = (req, res, next) => {
     let response = req.query.response || "";
 
     quizzes.findById(id)
-    .then((quiz) => res.send(play(id, quiz.question, response)))
-    .catch((error) => `A DB Error has occurred:\n${error}`);
- };
+        .then((quiz) => res.send(play(id, quiz.question, response)))
+        .catch((error) => `A DB Error has occurred:\n${error}`);
+};
 
 //  GET  /quizzes/1/check
 const checkController = (req, res, next) => {
@@ -139,69 +138,96 @@ const checkController = (req, res, next) => {
     let id = Number(req.params.id);
 
     quizzes.findById(id)
-    .then((quiz) => {
-        msg = (quiz.answer===response) ?
-              `Yes, "${response}" is the ${quiz.question}` 
-            : `No, "${response}" is not the ${quiz.question}`;
-        return res.send(check(id, msg, response));
-    })
-    .catch((error) => `A DB Error has occurred:\n${error}`);
+        .then((quiz) => {
+            msg = (quiz.answer === response) ?
+                `Yes, "${response}" is the ${quiz.question}`
+                : `No, "${response}" is not the ${quiz.question}`;
+            return res.send(check(id, msg, response));
+        })
+        .catch((error) => `A DB Error has occurred:\n${error}`);
 };
 
 //  GET /quizzes/1/edit
 const editController = (req, res, next) => {
 
-     // .... introducir código
+    // .... introducir código
+    let id = Number(req.params.id);
+
+    quizzes.findById(id)
+        .then((quiz) => {
+            return res.send(quizForm("Edit Quiz", "post", "Update", `/quizzes/${quiz.id}?_method=PUT`, quiz.question, quiz.answer));
+        })
+        .catch((error) => `A DB Error has occurred:\n${error}`);
 };
 
 //  PUT /quizzes/1
 const updateController = (req, res, next) => {
 
-     // .... introducir código
+    // .... introducir código
+
+    let id = Number(req.params.id);
+
+    quizzes.findById(id)
+        .then((quiz) => {
+            quiz.question=req.body.question;
+            quiz.answer=req.body.answer;
+            return quiz.save();
+        })
+        .then(() => indexController(req, res, next))
+        .catch((error) => `A DB Error has occurred:\n${error}`);
+
 };
 
 // GET /quizzes/new
 const newController = (req, res, next) => {
 
-    res.send(quizForm("Create new Quiz", "post", "/quizzes", "", ""));
- };
+    res.send(quizForm("Create new Quiz", "post", "Create", "/quizzes", "", ""));
+};
 
 // POST /quizzes
 const createController = (req, res, next) => {
     let {question, answer} = req.body;
 
     quizzes.build({question, answer})
-    .save()
-    .then((quiz) => res.redirect('/quizzes'))
-    .catch((error) => `Quiz not created:\n${error}`);
- };
+        .save()
+        .then((quiz) => res.redirect('/quizzes'))
+        .catch((error) => `Quiz not created:\n${error}`);
+};
 
 // DELETE /quizzes/1
 const destroyController = (req, res, next) => {
 
-     // .... introducir código
- };
+    // .... introducir código
+    console.log("he pasado por destroy");
+    let id = Number(req.params.id);
+    quizzes.destroy({where: {id}})
+        .then(() => indexController(req, res, next))
+        .catch((error) => `A DB Error has occurred:\n${error}`);
+};
 
 
+// ROUTER
 
-   // ROUTER
-
-app.get(['/', '/quizzes'],    indexController);
-app.get('/quizzes/:id/play',  playController);
+app.get(['/', '/quizzes'], indexController);
+app.get('/quizzes/:id/play', playController);
 app.get('/quizzes/:id/check', checkController);
-app.get('/quizzes/new',       newController);
-app.post('/quizzes',          createController);
+app.get('/quizzes/new', newController);
+app.post('/quizzes', createController);
 
-    // ..... instalar los MWs asociados a
-    //   GET  /quizzes/:id/edit,   PUT  /quizzes/:id y  DELETE  /quizzes/:id
+// ..... instalar los MWs asociados a
+//   GET  /quizzes/:id/edit,   PUT  /quizzes/:id y  DELETE  /quizzes/:id
+app.get('/quizzes/:id/edit', editController);
+app.put('/quizzes/:id', updateController);
+app.delete('/quizzes/:id', destroyController);
+
 
 
 app.all('*', (req, res) =>
     res.send("Error: resource not found or method not supported")
-);        
+);
 
 
-   // Server started at port 8000
+// Server started at port 8000
 
 app.listen(8000);
